refactor(artwork): move false-mode label+cache to data layer
Extract LOCAL_IMAGE=false helpers (filename→label parsing, LRU image
cache) from tools_artwork into data/artwork_mediawiki.{py,ts}, matching
the true-mode boundary (data/images ↔ tools_artwork). Pure code move,
no behavior change.
