+++ python/src/prts_mcp/tools_artwork.py
@@ -8,10 +8,8 @@ from __future__ import annotations
 import base64
 import json
 import logging
-import threading
-from collections import OrderedDict
 from pathlib import Path
-from typing import Annotated, Any, Literal, Mapping
+from typing import Annotated, Literal, Mapping
 
 from pydantic import Field
 
@@ -27,6 +25,12 @@ from prts_mcp.data.images import (
     _load_char_skins,
     build_artwork_label,
     parse_index,
+)
+from prts_mcp.data.artwork_mediawiki import (
+    _VARIANT_WIDTH,
+    _image_cache_get,
+    _image_cache_put,
+    _label_from_filename,
 )
 from prts_mcp.data.images_sync import _active_generation
 from prts_mcp.data.operator import _resolve_char_id
@@ -68,110 +72,6 @@ def _data_not_ready() -> object:
         "立绘数据未就绪。可能原因：IMAGES_ENABLED 未开启，或图片同步仍在进行中。"
         "请稍后重试；若持续不可用，请检查网络或 GITHUB_TOKEN。"
     )
-
-
-# ---------------------------------------------------------------------------
-# LOCAL_IMAGE=false MediaWiki path
-#
-# Data flows entirely from PRTS: allimages for file discovery, parsetree
-# (CharinfoV2 时装N名称) for fashion labels, imageinfo for variant URLs,
-# and download_image_safe for the pixel payload under the #85 boundary.
-# True (AKDP) and false (MediaWiki) modes share no data dependency.
-# ---------------------------------------------------------------------------
-
-_IMAGE_CACHE_LOCK = threading.Lock()
-_image_cache: "OrderedDict[str, bytes]" = OrderedDict()
-_IMAGE_CACHE_MAX_BYTES = 256 * 1024 * 1024  # 256 MiB (#85 §4.2)
-
-_MEDIAWIKI_BASE_LABELS: Mapping[str, str] = {"1": "精英零立绘", "2": "精英二立绘"}
-_VARIANT_WIDTH: Mapping[str, int] = {"large": 1024, "preview": 256}
-
-
-def _image_cache_key(artwork_id: str, variant: str) -> str:
-    return f"{artwork_id}|{variant}"
-
-
-def _image_cache_get(artwork_id: str, variant: str) -> bytes | None:
-    key = _image_cache_key(artwork_id, variant)
-    with _IMAGE_CACHE_LOCK:
-        if key in _image_cache:
-            _image_cache.move_to_end(key)
-            return _image_cache[key]
-    return None
-
-
-def _image_cache_put(artwork_id: str, variant: str, data: bytes) -> None:
-    key = _image_cache_key(artwork_id, variant)
-    with _IMAGE_CACHE_LOCK:
-        _image_cache[key] = data
-        _image_cache.move_to_end(key)
-        total = sum(len(v) for v in _image_cache.values())
-        while total > _IMAGE_CACHE_MAX_BYTES and _image_cache:
-            _, evicted = _image_cache.popitem(last=False)
-            total -= len(evicted)
-
-
-def _mediawiki_base_label(suffix: str) -> str:
-    base = suffix.rstrip("+")
-    plus = "+" in suffix
-    label = _MEDIAWIKI_BASE_LABELS.get(base)
-    if label is None:
-        label = f"立绘 {base}" if base else "立绘"
-    if plus:
-        label += "（变体）"
-    return label
-
-
-def _mediawiki_fashion_label(rest: str, charinfo: Mapping[str, Any]) -> str:
-    """rest is like 'skin3' (optionally 'skin2_sp'); leading digits = skin N."""
-    num = ""
-    for ch in rest[4:]:
-        if ch.isdigit():
-            num += ch
-        else:
-            break
-    label: str | None = None
-    if num:
-        val = charinfo.get(f"时装{num}名称")
-        if isinstance(val, str) and val:
-            label = val
-    if label is None:
-        label = f"时装 {num}" if num else "时装"
-    return label
-
-
-def _label_from_filename(
-    filename: str, charinfo: Mapping[str, Any],
-) -> str | None:
-    """Derive a label from a PRTS filename like ``立绘_阿米娅_2.png``.
-
-    Returns None for files we do not expose (建筑小人 ``_Nb``, non-png, or
-    names that don't match the ``立绘_<name>_<suffix>`` shape). Multi-form
-    operators carry the form in the name segment: ``立绘_阿米娅(近卫)_2.png``.
-    """
-    if not filename.endswith(".png"):
-        return None
-    base = filename[:-4]
-    parts = base.split("_", 2)
-    if len(parts) < 3:
-        return None
-    name = parts[1]
-    suffix = parts[2]
-    form: str | None = None
-    if "(" in name:
-        b = name.find("(")
-        e = name.find(")", b)
-        if 0 <= b < e:
-            form = name[b + 1:e]
-    if suffix.startswith("skin"):
-        label = _mediawiki_fashion_label(suffix, charinfo)
-    elif suffix.endswith("b"):
-        return None  # 建筑小人
-    else:
-        label = _mediawiki_base_label(suffix)
-    if form:
-        label += f"（{form}）"
-    return label
 
 
 async def _do_list_mediawiki(operator_name: str) -> object:
